+++ Raycoon.RayMigrator.Tests.Engine/Tests/MigrateUp/SqliteRollbackErrorOnlyTests.cs
@@ -28,7 +28,7 @@ public class SqliteRollbackErrorOnlyTests : SqliteTestBase
         await ctx.MigrateUpAsync();
 
         ctx.AssertSuccess(false);
-        ctx.AssertRunResult(MigrationRunResult.Error);
+        ctx.AssertRunResult(MigrationRunResult.Recovered);
         ctx.AssertRunCount(1);
 
         ctx.AssertFileStatuses(